Use of entrypoint in Dockerfile
... in order to not to need to put name in `docker run ...`

Now the output is not printed line-buffered, but full-buffered. That
means that buffer needs to be filled up to N bytes before be printed
out. It's the same effect as:

- `grep hola -`: grep is line-buffered, because it detects a console
output
- `grep hola - | head | grep hola`: Grep is full-buffered. You need to
fill the first grep buffer before data is sent to next pipe program,
same with head. This is because "grep detects that stdout is not a
terminal"
- `stdbuf -oL grep hola - | stdbuf -oL head | grep hola`: "Fix" full
buffered mode.